requirements.txt
google-generativeai
python-dotenv
streamlit
streamlit-login-auth-ui
argon2-cffi
os
json
streamlit_chat
dotenv
plotly
modules
datetime 
modules 
sklearn
PyPDF2
plotly
sentence_transformers
